crucible chages
